Setup modules for index.js
